+++ inst/examples/dungeonheroes/modules/realms/castle.R
@@ -16,8 +16,7 @@
     hide_map_navigation()
     game$activate_map(
       current_realm, player_name = "hero", x = 700, y = 500,
-      visible_objects = "blacksmith",
-      hidden_objects = c(mushroom_swamps_objects, wild_forests_objects, "talk_bubble_text")
+      objects = "blacksmith"
     )
     enemy_status_text$set("enemies: none in the castle")
     set_combat_status("The castle blacksmith is at his forge.")